+++ main.py
@@ -20,3 +20,4 @@ def baz():
   Prints out the text, "Hello world!"
   """
   print("Hello world!")
+
